requirements.txt
alembic==1.0.5
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Admin==1.5.3
Flask-Bootstrap==3.3.7.1
Flask-Excel==0.0.7
Flask-HTTPAuth==3.2.4
Flask-Login==0.4.1
flask-marshmallow==0.9.0
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
lml==0.0.9
Mako==1.0.7
MarkupSafe==1.1.0
marshmallow==2.17.0
marshmallow-sqlalchemy==0.15.0
mysqlclient==1.3.14
numpy==1.15.4
opencv-contrib-python==3.4.4.19
Pillow==5.3.0
pyexcel==0.5.10
pyexcel-io==0.5.11
pyexcel-webio==0.1.4
python-dateutil==2.7.5
python-editor==1.0.3
six==1.12.0
SQLAlchemy==1.2.15
texttable==1.5.0
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
XlsxWriter==1.1.2
